chore: replace Button to DialButton
