+++ apps/ai-dial-admin/src/components/Common/HintSIdebar/Hint.tsx
@@ -1,12 +1,10 @@
 'use client';
 
 import { FC } from 'react';
-import { IconX } from '@tabler/icons-react';
-import { DialButton } from '@epam/ai-dial-ui-kit';
 
-import { ButtonsI18nKey } from '@/src/constants/i18n';
 import { useAppContext } from '@/src/context/AppContext';
 import { useI18n } from '@/src/locales/client';
+import CloseButton from '@/src/components/Common/CloseButton/CloseButton';
 
 interface Props {
   title: string;
@@ -21,12 +19,7 @@ const Hint: FC<Props> = ({ title, text }) => {
     <div className="flex flex-col w-full gap-6">
       <div className="flex items-center justify-between">
         <h3 className="text-primary overflow-ellipsis">{title}</h3>
-        <DialButton
-          aria-label={t(ButtonsI18nKey.Close)}
-          cssClass="text-secondary hover:text-accent-primary"
-          onClick={closeHintSidebar}
-          iconBefore={<IconX height={24} width={24} />}
-        ></DialButton>
+        <CloseButton onClose={closeHintSidebar} />
       </div>
       <div className="overflow-y-scroll">
         <p className="small text-primary">{text}</p>
