+++ apps/ai-dial-admin/src/components/Grid/CellRenderers/KeyStatusCellRenderer.tsx
@@ -1,7 +1,8 @@
 'use client';
+
 import { ICellRendererParams } from 'ag-grid-community';
 
-import KeyViewStatus from '@/src/components/KeysList/KeyStatus/KeyViewStatus';
+import KeyViewStatus from '@/src/components/Keys/View/Header/KeyStatus';
 
 const KeyStatusCellRenderer = (params: ICellRendererParams) => {
   return <KeyViewStatus data={params.data} />;
