+++ apps/ai-dial-admin/src/components/Grid/CellRenderers/RadioButtonRenderer.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 
-import { DialRadioButton } from '@epam/ai-dial-ui-kit';
+import RadioButton from '@/src/components/Common/RadioButton/RadioButton';
 
 interface Props {
   inputId: string;
@@ -10,7 +10,7 @@ interface Props {
 const RadioButtonRenderer: FC<Props> = ({ inputId, isChecked }) => {
   return (
     <div className="h-6 w-6 flex items-center justify-center">
-      <DialRadioButton cssClass="w-[18px] h-[18px]" inputId={inputId} checked={isChecked} />
+      <RadioButton cssClass="w-[18px] h-[18px]" inputId={inputId} checked={isChecked} />
     </div>
   );
 };
